Refonte Model + relationships + fin auteurs + bougies

+++ app/controllers/BougiesController.php
@@ -3,7 +3,10 @@
 namespace app\controllers;
 
 use app\models\Bougie;
+use app\models\Collection;
+use app\models\Livre;
 use core\Controller;
+use core\Session;
 
 class BougiesController extends Controller
 {
@@ -17,60 +20,88 @@ class BougiesController extends Controller
 
     public function addForm()
     {
-
+        $this->view->render("bougies/add", [
+            'livres' => Livre::all(),
+            'collections' => Collection::all()
+        ]);
     }
 
     public function add()
     {
         $bougie = [
-            'nom_bougie' => "test ddddzad",
-            'id_livre' => 1,
-            'id_collection' => 1
+            'nom_bougie' => $this->request->post('name'),
+            'id_livre' => $this->request->post('id_livre'),
+            'id_collection' => $this->request->post('id_collection'),
+            'statut_bougie' => $this->request->post('statut_bougie')
         ];
 
-        $inserted = Bougie::create($bougie);
-
-        $this->view->render("bougies/show", [
-            'bougie' => $inserted
-        ]);
+        if (Bougie::unique($bougie['nom_bougie'], "nom_bougie"))
+        {
+            Bougie::create($bougie);
+            Session::addSuccess("Ajout réussi", "La bougie a bien été ajoutée.");
+            $this->redirect('/bougies');
+        }
+        else
+        {
+            Session::addError("Ajout impossible", "Cet bougie existe déjà");
+            Session::setOld(['name' => $bougie['nom_bougie']]);
+            $this->redirect('/bougies/add');
+        }
     }
 
     public function show($id)
     {
         $this->view->render("bougies/show", [
-            'bougie' => Bougie::find($id)
+            'bougie' => Bougie::findOrFail($id)
         ]);
     }
 
     public function updateForm($id)
     {
-        $this->view->render("bougies/edit", [
-            'bougie' => Bougie::find($id)
+        $this->view->render("bougies/update", [
+            'bougie' => Bougie::findOrFail($id),
+            'livres' => Livre::all(),
+            'collections' => Collection::all()
         ]);
     }
 
     public function update($id)
     {
         $bougie = [
-            'nom_bougie' => "Bougie de malade mental"
+            'nom_bougie' => $this->request->post('name'),
+            'id_livre' => $this->request->post('id_livre'),
+            'id_collection' => $this->request->post('id_collection'),
+            'statut_bougie' => $this->request->post('statut_bougie')
         ];
 
         Bougie::update($id, $bougie);
 
-        $this->redirect("/bougies/$id");
+        Session::addSuccess("Modification réussie", "La bougie a bien été modifiée.");
+
+        $this->redirect("/bougies");
     }
 
     public function deleteForm($id)
     {
         $this->view->render("bougies/delete", [
-            'bougie' => Bougie::find($id)
+            'bougie' => Bougie::findOrFail($id)
         ]);
     }
 
     public function delete($id)
     {
-        Bougie::delete($id);
+        $bougie = Bougie::findOrFail($id);
 
-        $this->redirect("/bougies");
+        if (count($bougie->recettes()) != 0)
+        {
+            Session::addError("Suppression impossible", "La bougie a encore des recettes.");
+        }
+        else
+        {
+            Bougie::delete($id);
+            Session::addSuccess("Suppression réussie", "La bougie a bien été supprimée.");
+        }
+
+        $this->redirect('/bougies');
     }
 }
